small refactoring
split code into modules
